Be more flexible to reuse, esp by factory classes.
_build_instance_with_traits is introduced so that a factory class which
composes MooseX::Traits::Pluggable can build an instance of an entirely
different class with traits applied.

If the default coderef for the _trait_namespace attribute is called it
is now passed the class or instance it belongs to as the first argument
(making this behave more like a normal moose default.

If a _trait_namespace method is present it will be called as usual
with arguments of the trait name being resolved.

Example of it being used is 8276ee in CatalystX::JobServer
http://goatse.co.uk/~bobtfish/Gitalist/script/gitalist.fcgi/CatalystX-JobServer.git/8276eec6e9f071fcbacc9cd9ecd938532a17a955/diff